// apps/fe-wtc-tech-test-angularv2/src/app/shared/models/movie.model.ts
import { Rating } from "./rating.model";

export interface Movie{
    Title: string;
    Year: string;
    Rated: string;
    Released: string;
    Runtime: string;
    Genre: string;
    Director: string;
    Writer: string;
    Actors: string;
    Plot: string;
    Language: string;
    Country: string;
    Awards: string;
    Poster: string;
    Ratings: Rating[];
    Metascore: string;
    imdbRating: string;
    imdbVotes: string;
    imdbID: string;
    Type: string;
    DVD: string;
    BoxOffice: string;
    Production: string;
    Website: string;
    Watched: boolean | string;
    Saved: boolean | string;
    finalRating?: string;
    stars: number[];
    background: string;
}